+++ src/routes/disambiguation.rs
@@ -14,10 +14,12 @@ struct IdPairQueryOptions {
 pub async fn to_msg(
     data: web::Data<ApplicationState<'_>>,
     query: web::Query<IdPairQueryOptions>,
-    ) -> impl Responder {
+) -> impl Responder {
     let client = data.db_client.lock().await;
-    if query.idpair.contains('.') { // if both message and submessage are specified
-        let parts: Vec<Result<i64, _>> = query.idpair.split('.').map(|x| x.parse::<i64>()).collect();
+    if query.idpair.contains('.') {
+        // if both message and submessage are specified
+        let parts: Vec<Result<i64, _>> =
+            query.idpair.split('.').map(|x| x.parse::<i64>()).collect();
         if parts[0].is_err() || parts[1].is_err() {
             return web::Redirect::to("/error?error_code=500").see_other();
         }
@@ -25,8 +27,9 @@ pub async fn to_msg(
         let parts: Vec<i64> = parts.into_iter().map(|x| x.unwrap()).collect();
         let msg = client.get_single_submessage(parts[0], parts[1]).await;
         match msg {
-            Ok(row) => return web::Redirect::to(format!("{}/topic/{}#{}", row.board, parts[0], parts[1])).see_other(),
-            Err(_) => return web::Redirect::to("/error?error_code=404").see_other(),
+            Ok(row) => web::Redirect::to(format!("{}/topic/{}#{}", row.board, parts[0], parts[1]))
+                .see_other(),
+            Err(_) => web::Redirect::to("/error?error_code=404").see_other(),
         }
     } else {
         let msgid = query.idpair.parse::<i64>();
@@ -36,8 +39,8 @@ pub async fn to_msg(
         let msgid = msgid.unwrap();
         let msg = client.get_single_message(msgid).await;
         match msg {
-            Ok(row) => return web::Redirect::to(format!("{}/topic/{}", row.board, msgid)).see_other(),
-            Err(_) => return web::Redirect::to("/error?error_code=404").see_other(),
+            Ok(row) => web::Redirect::to(format!("{}/topic/{}", row.board, msgid)).see_other(),
+            Err(_) => web::Redirect::to("/error?error_code=404").see_other(),
         }
     }
 }