+++ src/routes/report.rs
@@ -1,11 +1,20 @@
 //! Handler for message reporting
 
 use crate::routes::ApplicationState;
-use actix_web::{get, web, HttpResponse, Responder};
+use actix_web::{get, post, web, HttpResponse, Responder};
 
 /// Query params that specify reported messages
 #[derive(serde::Deserialize)]
 struct ReportQueryOptions {
+    id: i64,
+    subid: Option<i64>,
+}
+
+/// Form for report confirmation captcha
+#[derive(serde::Deserialize)]
+struct ReportCaptchaForm {
+    captcha_answer: String,
+    captcha_hash: String,
     id: i64,
     subid: Option<i64>,
 }
@@ -16,18 +25,47 @@ pub async fn report_msg(
     data: web::Data<ApplicationState<'_>>,
     page_data: web::Query<ReportQueryOptions>,
 ) -> impl Responder {
+    let captcha_hash = sha256::digest(crate::routes::create_new_captcha().await);
+    HttpResponse::Ok().body(
+        data.formatter
+            .format_into_report_captcha("".to_string(), captcha_hash, page_data.id, page_data.subid)
+            .await,
+    )
+}
+
+/// Handler for receiving captcha answer
+#[post("/report")]
+pub async fn report_process_captcha(
+    data: web::Data<ApplicationState<'_>>,
+    form: web::Form<ReportCaptchaForm>,
+) -> impl Responder {
+    if form.captcha_hash.clone() != sha256::digest(form.captcha_answer.clone()) {
+        return HttpResponse::Ok().body(
+            data.formatter
+                .format_into_error(actix_web::http::StatusCode::FORBIDDEN)
+                .await,
+        );
+    }
+
+    // delete captcha image after usage
+    // does not delete image if user got the captcha wrong... bug or feature? idk
+    if crate::routes::delete_captcha_image(form.captcha_answer.to_string())
+        .await
+        .is_err()
+    {
+        log::error!(
+            "Failed to delete used CAPTCHA: ./data/captcha/ACSIM_CAPTCHA_{}.png",
+            form.captcha_hash.to_string()
+        );
+    }
+
     let client = data.db_client.lock().await;
-    let message_type = match page_data.subid {
+    let message_type = match form.subid {
         Some(_v) => "submsg",
         None => "msg",
     };
     client
-        .insert_to_flagged(message_type.to_string(), page_data.id, page_data.subid)
+        .insert_to_flagged(message_type.to_string(), form.id, form.subid)
         .await;
-    let submsg_optional = match page_data.subid {
-        Some(n) => format!("#{}", n),
-        None => "".to_string(),
-    };
-    let backlink = format!("{}{}", page_data.id, submsg_optional);
-    HttpResponse::Ok().body(data.formatter.format_into_report(backlink).await)
+    HttpResponse::Ok().body(data.formatter.format_into_report_accepted().await)
 }